Added abstract BitcoinBasedBalanceUpdater with common logic for BTC and LTC. #99
